requirements.txt
numba
discord
py-cpuinfo
psutil
koreanbots